requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.3
numpy==1.23.3
openpyxl==3.0.10
pandas==1.4.4
pipreqs==0.4.11
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
yarg==0.1.9

PyMySQL~=1.0.2